+++ ai/openai/fake/src/main/kotlin/org/http4k/connect/openai/FakeOpenAI.kt
@@ -61,7 +61,7 @@ class FakeOpenAI(
     /**
      * Convenience function to get Ollama client
      */
-    fun client() = OpenAI.Http(OpenAIToken.of("openai-key"))
+    fun client() = OpenAI.Http(OpenAIToken.of("openai-key"), this)
 }
 
 fun main() {